requirements.txt
faiss-cpu>=1.7
sentence-transformers>=2.7
pymupdf>=1.24
requests>=2.31
numpy>=1.26
python-docx>=1.1
python-pptx>=0.6
pytesseract>=0.3
pillow>=10.0
